[libc] Fix some math conversion warnings
Differential Revision: https://reviews.llvm.org/D146738

+++ libc/src/math/generic/atanf.cpp
@@ -22,7 +22,7 @@ LLVM_LIBC_FUNCTION(float, atanf, (float x)) {
 
   if (LIBC_UNLIKELY(xbits.is_inf_or_nan())) {
     if (xbits.is_inf())
-      return opt_barrier(sign ? -M_MATH_PI_2 : M_MATH_PI_2);
+      return static_cast<float>(opt_barrier(sign ? -M_MATH_PI_2 : M_MATH_PI_2));
     else
       return x;
   }
@@ -52,7 +52,7 @@ LLVM_LIBC_FUNCTION(float, atanf, (float x)) {
     }
   }
 
-  return atan_eval(x);
+  return static_cast<float>(atan_eval(x));
 }
 
 } // namespace __llvm_libc
